🚚 rename LoggedInScene to be AuthScene
